Adds back map with randomized start locations.
Fixed start locations removed. Roughly 50% of the map remains as no-start locations due to either being too good or too bad. The criteria for eliminating a location as a potential start for land nations were locations with less then 5 connections and more then 6. Start locations needed a minimum of 5 connections to land provinces. Impassable mountain locations did not count toward this minimum, but did count toward the total number of connections for elimination.